backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
pypdf2==3.0.1
python-dotenv==1.0.0
reportlab==4.0.7
requests==2.31.0
google-genai==0.2.2
